requirements.txt
numpy==2.4.4
pillow==12.2.0
pyserial==3.5
PySide6==6.11.0
pywin32==311
WMI==1.5.1
scikit-learn>=0.24.1
scikit-image>=0.18.1